Users sql integer to BigInteger

+++ ErzaScarlet/modules/sql/afk_sql.py
@@ -1,13 +1,13 @@
 ﻿import threading
 
 from ErzaScarlet.modules.sql import BASE, SESSION
-from sqlalchemy import Boolean, Column, Integer, UnicodeText
+from sqlalchemy import Boolean, Column, Integer, UnicodeText,BigInteger
 
 
 class AFK(BASE):
     __tablename__ = "afk_users"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     is_afk = Column(Boolean)
     reason = Column(UnicodeText)
